requirements.txt
requests>=2.31.0
httpx>=0.26.0
pyyaml>=6.0
fastapi>=0.104.0
uvicorn>=0.24.0

# Optional: Advanced WAF bypass (requires venv)
# curl-cffi>=0.6.0
